Move get_block_locators to snarkos-node-sync-locators
